% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalCIRHist}
\alias{CalCIRHist}
\title{CIR Historical Calibration}
\usage{
CalCIRHist(
  n_years = 99,
  to = NULL,
  from = NULL,
  shift = NULL,
  return.data = FALSE
)
}
\arguments{
\item{return.data}{}
}
\value{

}
\description{
CIR Historical Calibration
}
