% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalEquityILN}
\alias{CalEquityILN}
\title{Calibrate Equity Returns using}
\usage{
CalEquityILN(
  n_years = 20,
  from = NULL,
  to = NULL,
  return.data = FALSE,
  symbol = "^GSPC"
)
}
\arguments{
\item{n_years}{number of historical years to include in calibration}

\item{to}{a character string representing a end date in YYYY-MM-DD format}

\item{return.data}{}
}
\value{
list of ML parameters and data
}
\description{
Get Historical mean and standard deviation of monthly log returns
}
