% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalILN}
\alias{CalILN}
\title{Calibrate Independent LogNormal Parameters}
\usage{
CalILN(dat, dt = 1/12)
}
\arguments{
\item{dat}{numeric vector of log returns}

\item{dt}{time step, default is monthly (1/12)}
}
\value{
vector with annual (arithmetic) mean and stdev of log returns
}
\description{
Get Historical mean and standard deviation of log returns
}
