% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalInflation}
\alias{CalInflation}
\title{Calibrate Inflation}
\usage{
CalInflation(n_years = 20, from = NULL, to = NULL, return.data = FALSE)
}
\arguments{
\item{n_years}{number of historical years to include in calibration}

\item{to}{a character string representing a end date in YYYY-MM-DD format}

\item{return.data}{}
}
\value{
list of ML parameters and data
}
\description{
Get Historical Inflation and estimate parameters of 1-factor Vasicek Model
}
