% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalVasicekHist}
\alias{CalVasicekHist}
\title{Vasicek Historical Calibration}
\usage{
CalVasicekHist(n_years = 99, to = NULL, from = NULL, return.data = F)
}
\arguments{
\item{dt}{}
}
\value{

}
\description{
Calibrates the vasicek model using the maximum likelihood estimator
}
