% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\docType{data}
\name{cas_inflation_vas1f}
\alias{cas_inflation_vas1f}
\title{Default Inflation Parameters (CAS-SOA)}
\format{
An object of class \code{list} of length 5.
}
\usage{
cas_inflation_vas1f
}
\description{
List of stochastic process parameters
}
\details{
Convenience list of stochastic process parameters for Vasicek Model
}
\keyword{datasets}
