% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\docType{data}
\name{cas_rates_vas2f}
\alias{cas_rates_vas2f}
\title{Default 2 factor Vasicek Real Interest Rate Parameters (CAS-SOA)}
\format{
An object of class \code{list} of length 2.
}
\usage{
cas_rates_vas2f
}
\description{
List of stochastic process parameters
}
\details{
Convenience list of stochastic process parameters for two-factor Vasicek Model
}
\keyword{datasets}
