% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreatePWL}
\alias{CreatePWL}
\title{Create Piecewise Linear from Simulation}
\usage{
CreatePWL(dat, bins = 100, xmin = NULL, xmax = NULL, remove.dup = TRUE)
}
\arguments{
\item{dat}{numeric vector conatining simulation data}

\item{bins}{optional number of bins}

\item{xmin}{optional minimum value}

\item{xmax}{optional maximum value}

\item{remove.dup}{optional argument to remove bins with no probability mass}
}
\value{
tibble with columns x and Fx
}
\description{
Create Piecewise Linear from Simulation
}
