% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-equity.R
\name{.GetStates}
\alias{.GetStates}
\title{Get states vector of regime switching process}
\usage{
.GetStates(n = 1, t = 12, pswitch = c(0.011, 0.059), state0 = c(0, 1))
}
\arguments{
\item{n}{number of simulated trials}

\item{t}{number of future monthly timesteps}

\item{pswitch}{vector of length two with transition probabilities for each state}

\item{state0}{Optional intial state. If unchanged will simulate from unconditional state probabilities}
}
\value{

}
\description{
Get states vector of regime switching process
}
\keyword{internal}
