% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-rates.R
\name{.SimDiffusion}
\alias{.SimDiffusion}
\title{Simulate mean reverting diffusion process}
\usage{
.SimDiffusion(
  n = 1,
  param_list = list(r0, a, b, v),
  rmin = NULL,
  t = 12,
  g = 0,
  z_rand = NULL,
  col_rename = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{number of simulated trials}

\item{param_list}{a named list with parameters r0 - initial value,
a - speed of mean reversion (annual), b - mean reversion level,  v - annual volatility}

\item{rmin}{\itemize{
\item optional rate lower bound
}}

\item{t}{number of future monthly timesteps}

\item{g}{optional proportional conditional volatility exponent (0 = Vasicek, 1/2 = CIR)}

\item{z_rand}{optional vector standard normal random variables of length n * t}

\item{col_rename}{optional name of output column}
}
\value{
data in the form of a \code{tibble} object with columns trial, time, r (or col_rename)
}
\description{
Monthly Simulation of rates using Ornstein-Uhlenbeck Model
}
\keyword{internal}
