% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-curve.R
\name{.Vas2FA}
\alias{.Vas2FA}
\title{Calculate A parameter of 2 factor Vasicek Model}
\usage{
.Vas2FA(a1, a2, s1, s2, u, s)
}
\description{
Calculate A parameter of 2 factor Vasicek Model
}
\keyword{internal}
