% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equity}
\alias{equity}
\title{Historical S&P500 Monthly Returns}
\format{
A data frame with 2 variables:
\describe{
\item{date}{Calendar Month}
\item{return}{log return}
...
}
}
\source{
\url{https://finance.yahoo.com/quote/\%5EGSPC}
}
\usage{
equity
}
\description{
A dataset containing historical log (continous) returns of the S&P500 index.
}
\keyword{datasets}
