% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetTreasuries}
\alias{GetTreasuries}
\title{Get Treasury Yield Curve as of given date}
\usage{
GetTreasuries(date = NULL)
}
\arguments{
\item{date}{a character string representing a date in YYYY-MM-DD format}
}
\value{
tibble with columns date, symbol, yield
}
\description{
Get Treasury Yield Curve as of given date
}
\examples{
GetTreasuries("2022-02-04")
}
