% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gg_funnel}
\alias{gg_funnel}
\title{Funnel Uncertainty Plot}
\usage{
gg_funnel(
  data_forecast,
  data_actual,
  init_value,
  variable_name = NULL,
  time_offset = 0
)
}
\arguments{
\item{data_forecast}{tibble with columns time, value}

\item{data_actual}{numeric vector of historical monthly}

\item{init_value}{value at time 0}

\item{variable_name}{A string or character vector containing name of projected variable}

\item{time_offset}{numeric offset for x-axis (e.g. time 0 = 2022)}
}
\value{
ggplot object
}
\description{
Funnel Uncertainty Plot
}
