% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_rates}
\alias{long_rates}
\title{Historical Yield on U.S. Treasury Securities at 10-Year Constant Maturity}
\format{
A data frame with 2 variables:
\describe{
\item{date}{Calendar Month}
\item{nominal}{nominal yield}
\item{real}{real yield estimated as nominal less monthly inflation}
...
}
}
\source{
\url{https://fred.stlouisfed.org/series/DGS10}
}
\usage{
long_rates
}
\description{
A dataset containing historical 10-year treasury yields (nominal)
}
\keyword{datasets}
