% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quibble}
\alias{quibble}
\title{Quantile Summary of distribution}
\usage{
quibble(
  x,
  q = c(0.005, 0.01, 0.025, 0.05, 0.1, 0.5, 0.9, 0.95, 0.975, 0.99, 0.995)
)
}
\arguments{
\item{x}{data to be summarized}

\item{q}{vector of quantiles}
}
\value{
tibble with mean, standard deviation and quantiles
}
\description{
Quantile Summary of distribution
}
