% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_rates}
\alias{short_rates}
\title{Historical Yield on U.S. Treasury Securities at 3-Month Constant Maturity}
\format{
A data frame with 484 rows and 2 variables:
\describe{
\item{date}{Calendar Month}
\item{yield}{yield}
...
}
}
\source{
\url{https://fred.stlouisfed.org/series/DGS3MO}
}
\usage{
short_rates
}
\description{
A dataset containing historical seasonally adjusted CPI and inflation.
}
\keyword{datasets}
