% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-rates.R
\name{SimCIR1F}
\alias{SimCIR1F}
\title{Simulate 1 factor Cox Ingersoll Ross (CIR) model}
\usage{
SimCIR1F(
  n = 1,
  param_list = list(r0 = 0.01, a = 0.5, b = 0.05, v = 0.05),
  rmin = NULL,
  t = 12,
  z_rand = NULL,
  col_rename = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{number of simulated trials}

\item{param_list}{a list with parameters r0 - initial value,
a - speed of mean reversion (annual), b - mean reversion level, v - annual volatility}

\item{rmin}{\itemize{
\item optional rate lower bound
}}

\item{t}{number of future monthly timesteps}

\item{z_rand}{optional vector standard normal random variables of length n * t}

\item{col_rename}{optional name of output column}

\item{seed}{optional starting seed}

\item{shift}{optional additive shift post simulation}
}
\value{
data in the form of a \code{tibble} object with columns trial, time, r (or col_rename)
}
\description{
Monthly Simulation of rates using 1 factor CIR model
}
\examples{
SimCIR1F(param_list = list(r0=.01, a=.5, b=.05, v=.05))
}
