% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-equity.R
\name{SimEquityILN}
\alias{SimEquityILN}
\title{Simulate Independent Lognormal (ILN)}
\usage{
SimEquityILN(
  n = 1,
  param = c(0.08, 0.16),
  t = 12,
  type = c("arithmetic", "log"),
  seed = NULL
)
}
\arguments{
\item{n}{number of simulated trials}

\item{param}{Vector of length two containing \link{1} - expected annual (arithmetic) return
\link{2} volatility of stock price per year (sigma of log returns)}

\item{t}{number of future monthly timesteps}

\item{type}{type of output returns: arithmetic (discrete) or log (continuous)}

\item{seed}{optional starting seed}
}
\value{
data in the form of a \code{tibble} object with columns trial, time, mo_return, cum_return, accum
}
\description{
Simulate monthly returns assuming log returns in non-overlapping are iid
normal with constant mean and volatility
}
\examples{
SimEquityILN(n=10, c(.08, .16))
}
