% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-equity.R
\name{SimEquityRSLN}
\alias{SimEquityRSLN}
\title{Regime Switching LogNormal Equity Return Model}
\usage{
SimEquityRSLN(
  n = 1,
  pswitch = c(0.011, 0.059),
  means = c(0.008, -0.011),
  vols = c(0.039, 0.113),
  t = 12,
  state0 = c(0, 1),
  mean_uncond = 0,
  z_rand = NULL,
  type = c("arithmetic", "log"),
  .detail = FALSE
)
}
\arguments{
\item{n}{number of simulated trials}

\item{pswitch}{vector of length two with transition probabilities for each state}

\item{means}{vector of length two with monthly mean returns conditional on state
(added to optional unconditional mean)}

\item{vols}{vector of length two with monthly volatility conditional on state}

\item{t}{number of future monthly timesteps}

\item{state0}{Optional intial state. If unchanged will simulate from unconditional state probabilities}

\item{mean_uncond}{Optional vector of length (n * t) additional ANNUAL return unconditional on state (e.g. real return + inflation)}

\item{z_rand}{Optional normal standard deviate vector of length (n * t)}

\item{type}{type of output returns: arithmetic (discrete) or log (continuous)}

\item{.detail}{Include additional detail state, mean, volatility}
}
\value{
data in the form of a \code{tibble} object with columns trial, time, mo_return, cum_return, accum
}
\description{
Simulate monthly equity returns from a 2-state RSLN model
}
\examples{
SimEquityRSLN(n=10, pswitch=c(.011, .059), means=c(.008, -.011), vols = c(.039, .113))
}
