% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-rates.R
\name{SimVasicek2F}
\alias{SimVasicek2F}
\title{Simulate 2 factor Vasicek Model (stochastic mean reverting term)}
\usage{
SimVasicek2F(
  n = 1,
  param_short = list(r0 = 0.01, m = 0.05, k = 0.5, v = 0.05, rmin = NULL),
  param_long = list(r0 = 0.01, m = 0.05, k = 0.5, v = 0.05, rmin = NULL),
  t = 12,
  rcorr = 0.5,
  seed = NULL
)
}
\arguments{
\item{n}{number of simulated trials}

\item{param_short}{(short rate ) a named list with parameters r0 - initial value,
m - mean reversion level, k - speed of mean reversion (annual), v - annual volatility, rmin - optional rate lower bound}

\item{param_long}{(mean reversion term) a named list with parameters r0 - initial value,
m - mean reversion level, k - speed of mean reversion (annual), v - annual volatility, rmin - optional rate lower bound}

\item{t}{number of future monthly timesteps}

\item{rcorr}{correlation between long and short processes}

\item{seed}{optional starting seed}
}
\value{
data in the form of a \code{tibble} object with columns trial, time, r1, r2
}
\description{
Monthly Simulation of Inflation using 2-factor Vasicek (Hull White) Model
}
\examples{
SimVasicek2F(param_short = list(r0=.01, m=.05, k=.5, v=.05), param_long = list(r0=.01, m=.05, k=.5, v=.05))
}
