% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-curve.R
\name{Vas2FBondYield}
\alias{Vas2FBondYield}
\title{Bond Yield using 2 factor Vasicek Model}
\usage{
Vas2FBondYield(r1, r2, a1, a2, s1, s2, u, t)
}
\arguments{
\item{t}{}
}
\value{

}
\description{
Bond Yield using 2 factor Vasicek Model
}
