% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-calcs.R
\name{Vasicek1fPrice}
\alias{Vasicek1fPrice}
\title{Price of bond under 1-factor Vasicek Model}
\usage{
Vasicek1fPrice(r, a, b, s, t)
}
\arguments{
\item{t}{}
}
\value{

}
\description{
Price of bond under 1-factor Vasicek Model
}
