% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-calcs.R, R/yield-curve.R
\name{VasicekYield}
\alias{VasicekYield}
\title{Bond Yield under 1-factor Vasicek Model}
\usage{
VasicekYield(r, a, b, s, t)

VasicekYield(r, a, b, s, t)
}
\arguments{
\item{t}{}
}
\value{



}
\description{
Bond Yield under 1-factor Vasicek Model

Bond Yield under Vasicek Model
}
