% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{CalVasicek1f}
\alias{CalVasicek1f}
\title{Calibrate 1-factor Vasicek Model using the Maximum Likelihood Estimator}
\usage{
CalVasicek1f(dat, dt = 1/12)
}
\arguments{
\item{dt}{time step, default is monthly (1/12)}

\item{data}{numerical vector of time series data}
}
\value{
a named list with parameters r0 - initial value,
a - speed of mean reversion (annual), b - mean reversion level, v - annual volatility
}
\description{
Calibrate 1-factor Vasicek Model using the Maximum Likelihood Estimator
}
