% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Historical CPI & Inflation}
\format{
A data frame with 888 rows and 3 variables:
\describe{
\item{date}{Calendar Month}
\item{price}{CPI}
\item{inf}{YoY \% change in CPI}
...
}
}
\source{
\url{https://fred.stlouisfed.org/series/CPIAUCSL}
}
\usage{
inflation
}
\description{
A dataset containing historical seasonally adjusted CPI and inflation.
}
\keyword{datasets}
