% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-calcs.R
\name{Vasicek2fYield}
\alias{Vasicek2fYield}
\title{Bond Yield under 2-factor Vasicek Model}
\usage{
Vasicek2fYield(r1, r2, a1, a2, s1, s2, u, t)
}
\arguments{
\item{t}{}
}
\value{

}
\description{
Bond Yield under 2-factor Vasicek Model
}
